remove prediction tests from model tests
