+++ tests/test_models.py
@@ -3,21 +3,17 @@
 """Test that models can be executed."""
 
 import importlib
-import itertools
 import os
 import unittest
-from typing import Any, Iterable, MutableMapping, Optional, Set, Type, Union
+from typing import Any, Iterable, MutableMapping, Set, Type, Union
 
-import numpy
 import torch
 import unittest_templates
 
 import pykeen.experiments
 import pykeen.models
-from pykeen.datasets.nations import Nations
 from pykeen.models import ERModel, EvaluationOnlyModel, FixedModel, Model, _NewAbstractModel, model_resolver
 from pykeen.models.multimodal.base import LiteralModel
-from pykeen.predict import get_all_prediction, predict_all
 from pykeen.nn import Embedding, NodePieceRepresentation
 from pykeen.nn.combination import ConcatAggregationCombination
 from pykeen.nn.perceptron import ConcatMLP
@@ -84,10 +80,6 @@ class TestConvE(cases.ModelTestCase):
     #                                 7
     num_constant_init = 7
 
-    def test_predict(self):
-        """Test prediction workflow with inverse relations."""
-        predict_all(model=self.instance, k=10)
-
 
 class TestConvKB(cases.ModelTestCase):
     """Test the ConvKB model."""
@@ -112,52 +104,6 @@ class TestDistMult(cases.ModelTestCase):
         """
         entity_norms = self.instance.entity_representations[0](indices=None).norm(p=2, dim=-1)
         assert torch.allclose(entity_norms, torch.ones_like(entity_norms))
-
-    def _test_score_all_triples(self, k: Optional[int], batch_size: int = 16):
-        """Test score_all_triples.
-
-        :param k: The number of triples to return. Set to None, to keep all.
-        :param batch_size: The batch size to use for calculating scores.
-        """
-        top_triples, top_scores = predict_all(model=self.instance, batch_size=batch_size, k=k)
-
-        # check type
-        assert torch.is_tensor(top_triples)
-        assert torch.is_tensor(top_scores)
-        assert top_triples.dtype == torch.long
-        assert top_scores.dtype == torch.float32
-
-        # check shape
-        actual_k, n_cols = top_triples.shape
-        assert n_cols == 3
-        if k is None:
-            assert actual_k == self.factory.num_entities**2 * self.factory.num_relations
-        else:
-            assert actual_k == min(k, self.factory.num_triples)
-        assert top_scores.shape == (actual_k,)
-
-        # check ID ranges
-        assert (top_triples >= 0).all()
-        assert top_triples[:, [0, 2]].max() < self.instance.num_entities
-        assert top_triples[:, 1].max() < self.instance.num_relations
-
-    def test_score_all_triples(self):
-        """Test score_all_triples with a large batch size."""
-        # this is only done in one of the models
-        self._test_score_all_triples(k=15, batch_size=16)
-
-    def test_score_all_triples_singleton_batch(self):
-        """Test score_all_triples with a batch size of 1."""
-        self._test_score_all_triples(k=15, batch_size=1)
-
-    def test_score_all_triples_large_batch(self):
-        """Test score_all_triples with a batch size larger than k."""
-        self._test_score_all_triples(k=10, batch_size=16)
-
-    def test_score_all_triples_keep_all(self):
-        """Test score_all_triples with k=None."""
-        # this is only done in one of the models
-        self._test_score_all_triples(k=None)
 
 
 class TestDistMA(cases.ModelTestCase):
@@ -583,24 +529,6 @@ class TestTransE(cases.DistanceModelTestCase):
         entity_norms = self.instance.entity_representations[0](indices=None).norm(p=2, dim=-1)
         assert torch.allclose(entity_norms, torch.ones_like(entity_norms))
 
-    def test_get_all_prediction_df(self):
-        """Test consistency of top-k scoring."""
-        ks = [5, 10]
-        dfs = [
-            get_all_prediction(
-                model=self.instance,
-                triples_factory=self.factory,
-                batch_size=1,
-                k=k,
-            )
-            .nlargest(n=min(ks), columns="score")
-            .reset_index(drop=True)
-            for k in ks
-        ]
-        assert set(dfs[0].columns) == set(dfs[0].columns)
-        for column in dfs[0].columns:
-            numpy.testing.assert_equal(dfs[0][column].values, dfs[1][column].values)
-
 
 class TestTransF(cases.ModelTestCase):
     """Test the TransF model."""
@@ -816,71 +744,6 @@ class ERModelTests(cases.ModelTestCase):
 
     def test_has_hpo_defaults(self):  # noqa: D102
         raise unittest.SkipTest(f"Base class {self.cls} does not provide HPO defaults.")
-
-
-class InverseRelationPredictionTests(unittest_templates.GenericTestCase[pykeen.models.FixedModel]):
-    """Test for prediction with inverse relations."""
-
-    cls = pykeen.models.FixedModel
-
-    def _pre_instantiation_hook(self, kwargs: MutableMapping[str, Any]) -> MutableMapping[str, Any]:
-        # create triples factory with inverse relations
-        kwargs = super()._pre_instantiation_hook(kwargs=kwargs)
-        kwargs["triples_factory"] = self.factory = Nations(create_inverse_triples=True).training
-        return kwargs
-
-    def _combination_batch(
-        self,
-        heads: bool = True,
-        relations: bool = True,
-        tails: bool = True,
-    ) -> torch.LongTensor:
-        """Generate a batch with all combinations."""
-        factors = []
-        if heads:
-            factors.append(range(self.factory.num_entities))
-        if relations:
-            factors.append(range(self.factory.real_num_relations))
-        if tails:
-            factors.append(range(self.factory.num_entities))
-        return torch.as_tensor(
-            data=list(itertools.product(*factors)),
-            dtype=torch.long,
-        )
-
-    def test_predict_hrt(self):
-        """Test predict_hrt."""
-        hrt_batch = self._combination_batch()
-        expected_scores = self.instance._generate_fake_scores(
-            h=hrt_batch[:, 0],
-            r=2 * hrt_batch[:, 1],
-            t=hrt_batch[:, 2],
-        ).unsqueeze(dim=-1)
-        scores = self.instance.predict_hrt(hrt_batch=hrt_batch)
-        assert torch.allclose(scores, expected_scores)
-
-    def test_predict_h(self):
-        """Test predict_h."""
-        rt_batch = self._combination_batch(heads=False)
-        # head prediction via inverse tail prediction
-        expected_scores = self.instance._generate_fake_scores(
-            h=rt_batch[:, 1, None],
-            r=2 * rt_batch[:, 0, None] + 1,
-            t=torch.arange(self.factory.num_entities).unsqueeze(dim=0),
-        )
-        scores = self.instance.predict_h(rt_batch=rt_batch)
-        assert torch.allclose(scores, expected_scores)
-
-    def test_predict_t(self):
-        """Test predict_t."""
-        hr_batch = self._combination_batch(tails=False)
-        expected_scores = self.instance._generate_fake_scores(
-            h=hr_batch[:, 0, None],
-            r=2 * hr_batch[:, 1, None],
-            t=torch.arange(self.factory.num_entities).unsqueeze(dim=0),
-        )
-        scores = self.instance.predict_t(hr_batch=hr_batch)
-        assert torch.allclose(scores, expected_scores)
 
 
 class CooccurrenceFilteredModelTests(cases.ModelTestCase):
